move DynamicColumn to separate header
