+++ Core/include/Acts/EventData/VectorMultiTrajectory.hpp
@@ -11,6 +11,7 @@
 #include "Acts/Definitions/TrackParametrization.hpp"
 #include "Acts/EventData/MultiTrajectory.hpp"
 #include "Acts/EventData/TrackStatePropMask.hpp"
+#include "Acts/EventData/detail/DynamicColumn.hpp"
 
 #include <unordered_map>
 
@@ -23,72 +24,6 @@ namespace detail_vmt {
 using MultiTrajectoryTraits::IndexType;
 constexpr auto kInvalid = MultiTrajectoryTraits::kInvalid;
 constexpr auto MeasurementSizeMax = MultiTrajectoryTraits::MeasurementSizeMax;
-
-struct DynamicColumnBase {
-  virtual ~DynamicColumnBase() = 0;
-
-  virtual std::any get(size_t i) = 0;
-  virtual std::any get(size_t i) const = 0;
-
-  virtual void add() = 0;
-  virtual void clear() = 0;
-
-  virtual std::unique_ptr<DynamicColumnBase> clone() const = 0;
-};
-
-inline DynamicColumnBase::~DynamicColumnBase() = default;
-
-template <typename T>
-struct DynamicColumn : public DynamicColumnBase {
-  ~DynamicColumn() override = default;
-
-  std::any get(size_t i) override {
-    assert(i < m_vector.size() && "DynamicColumn out of bounds");
-    return &m_vector[i];
-  }
-
-  std::any get(size_t i) const override {
-    assert(i < m_vector.size() && "DynamicColumn out of bounds");
-    return &m_vector[i];
-  }
-
-  void add() override { m_vector.emplace_back(); }
-  void clear() override { m_vector.clear(); }
-
-  std::unique_ptr<DynamicColumnBase> clone() const override {
-    return std::make_unique<DynamicColumn<T>>(*this);
-  }
-
-  std::vector<T> m_vector;
-};
-
-template <>
-struct DynamicColumn<bool> : public DynamicColumnBase {
-  struct Wrapper {
-    bool value;
-  };
-
-  ~DynamicColumn() override = default;
-
-  std::any get(size_t i) override {
-    assert(i < m_vector.size() && "DynamicColumn out of bounds");
-    return &m_vector[i].value;
-  }
-
-  std::any get(size_t i) const override {
-    assert(i < m_vector.size() && "DynamicColumn out of bounds");
-    return &m_vector[i].value;
-  }
-
-  void add() override { m_vector.emplace_back(); }
-  void clear() override { m_vector.clear(); }
-
-  std::unique_ptr<DynamicColumnBase> clone() const override {
-    return std::make_unique<DynamicColumn<bool>>(*this);
-  }
-
-  std::vector<Wrapper> m_vector;
-};
 
 class VectorMultiTrajectoryBase {
  public:
@@ -312,8 +247,8 @@ class VectorMultiTrajectoryBase {
         if (it == instance.m_dynamic.end()) {
           throw std::runtime_error("Unable to handle this component");
         }
-        std::conditional_t<EnsureConst, const DynamicColumnBase*,
-                           DynamicColumnBase*>
+        std::conditional_t<EnsureConst, const detail::DynamicColumnBase*,
+                           detail::DynamicColumnBase*>
             col = it->second.get();
         assert(col && "Dynamic column is null");
         return col->get(istate);
@@ -369,7 +304,7 @@ class VectorMultiTrajectoryBase {
   // be handled in a smart way by moving but not sure.
   std::vector<std::shared_ptr<const Surface>> m_referenceSurfaces;
 
-  std::unordered_map<HashedString, std::unique_ptr<DynamicColumnBase>>
+  std::unordered_map<HashedString, std::unique_ptr<detail::DynamicColumnBase>>
       m_dynamic;
 };
 
@@ -479,7 +414,7 @@ class VectorMultiTrajectory final
   template <typename T>
   constexpr void addColumn_impl(const std::string& key) {
     m_dynamic.insert(
-        {hashString(key), std::make_unique<detail_vmt::DynamicColumn<T>>()});
+        {hashString(key), std::make_unique<detail::DynamicColumn<T>>()});
   }
 
   constexpr bool hasColumn_impl(HashedString key) const {
